reader: do u32 -> u64 conversion higher up the stack
we're about to implement a file reader which will use this method. since
the u32 bits are a result of fuse, let's move this a bit higher up the
stack so we have to do less casting in that new file reader.

+++ reader/src/fuse.rs
@@ -81,7 +81,7 @@ impl<'a> Fuse<'a> {
 
     fn _read(&mut self, ino: u64, offset: u64, size: u32) -> FSResult<Vec<u8>> {
         let inode = self.pfs.find_inode(ino)?;
-        self.pfs.file_read(&inode, offset, size)
+        self.pfs.file_read(&inode, offset, u64::from(size))
     }
 
     fn _readdir(
